对各个controller进行了整合，连接了前端首页的初始化

+++ src/main/java/com/uestc2h/u2pc/service/UserService.java
@@ -6,7 +6,7 @@ public interface UserService {
 
     UserDTO login(UserDTO user);
 
-    int register(UserDTO user);
+    Long register(UserDTO user);
 
     int setInfo(UserDTO user);
 }
